pyproject.toml
[tool.poetry]
name = "cerebras-bank-analyzer"
version = "0.1.0"
description = "AI-powered bank statement analysis using Cerebras and OpenRouter"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pdfplumber = "^0.10.4"
pandas = "^2.1.3"
openpyxl = "^3.1.2"
supabase = "^2.4.5"
stripe = "^8.10.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pydantic = "^2.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"